Agent_01_Financial_Analyst/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
googlesearch-python
pycountry